last gigit

+++ 0x01-variables_if_else_while/1-last_digit.c
@@ -14,12 +14,12 @@ int main(void)
 int n;
 srand(time(0));
 n = rand() - RAND_MAX / 2;
-int lstdgt = n % 10;
+long lstdgt = n % 10;
 if (lstdgt == 0)
-printf("Last digit of %d is %d and is 0\n", n, lstdgt);
+printf("Last digit of %d is %ld and is 0\n", n, lstdgt);
 else if (lstdgt < 6)
-printf("Last digit of %d is %d and is less than 6 not 0\n", n, lstdgt);
+printf("Last digit of %d is %ld and is less than 6 not 0\n", n, lstdgt);
 else
-printf("Last digit of %d is %d an is greater than 5\n", n, lstdgt);
+printf("Last digit of %d is %ld an is greater than 5\n", n, lstdgt);
 return (0);
 }
